move logging settings handling to separate file
