+++ logging/logger.go
@@ -2,30 +2,11 @@ package logging
 
 import (
 	"context"
-	"io"
-	"os"
-	"strings"
 	"sync"
 	"time"
 
 	"github.com/rs/zerolog"
 )
-
-// Config holds the settings for the logger
-type Config struct {
-
-	// Level is the log Level for the logger that will be used
-	Level zerolog.Level `yaml:"log_level" env:"LOG_LEVEL" env-default:"info"`
-
-	// Format is the log Format for the logger that will be used
-	Format LogFormat `yaml:"log_format" env:"LOG_FORMAT" env-default:"text"`
-
-	// output is the name of the output for the logger that will be used
-	output LogOutput `yaml:"log_output" env:"LOG_OUTPUT" env-default:"stdout"`
-
-	// Output is the Output for the logger that will be used
-	Output io.Writer
-}
 
 var (
 	defaultLogger zerolog.Logger
@@ -60,10 +41,10 @@ func Init(config Config) {
 
 // With returns a new context with the provided logger or with the default logger if no logger is provided
 func With(ctx context.Context, logger ...zerolog.Logger) context.Context {
-	if len(logger) > 0 {
-		return logger[0].WithContext(ctx)
+	if len(logger) == 0 {
+		return defaultLogger.WithContext(ctx)
 	}
-	return defaultLogger.WithContext(ctx)
+	return logger[0].WithContext(ctx)
 }
 
 // Get returns the logger from the context
@@ -79,75 +60,4 @@ func Get(ctx ...context.Context) *zerolog.Logger {
 	}
 
 	return logger
-}
-
-type LogFormat string
-
-const (
-	JsonFormat LogFormat = "json"
-	TextFormat LogFormat = "text"
-	TimeFormat string    = "2006-01-02T15:04:05.000000"
-)
-
-func (f LogFormat) String() string {
-	switch f {
-	case JsonFormat:
-		return "json"
-	default:
-		return "text"
-	}
-}
-
-func (f *LogFormat) SetValue(s string) error {
-	*f = ParseLogFormat(s)
-	return nil
-}
-
-func ParseLogFormat(s string) LogFormat {
-	switch strings.ToLower(s) {
-	case "json":
-		return JsonFormat
-	default:
-		return TextFormat
-	}
-}
-
-type LogOutput string
-
-const (
-	Stdout LogOutput = "stdout"
-	Stderr LogOutput = "stderr"
-)
-
-func (o LogOutput) String() string {
-	switch o {
-	case Stdout:
-		return "stdout"
-	case Stderr:
-		return "stderr"
-	default:
-		return "stdout"
-	}
-}
-
-func ParseLogOutput(s string) LogOutput {
-	switch strings.ToLower(s) {
-	case "stdout":
-		return Stdout
-	case "stderr":
-		return Stderr
-	default:
-		return Stdout
-	}
-}
-
-func getOutput(o LogOutput) io.Writer {
-	switch o {
-	case Stdout:
-		return os.Stdout
-	case Stderr:
-		return os.Stderr
-	default:
-		return os.Stdout
-	}
 }
